Extract common API client mock server class
